Almost completed

+++ MovieManagement.Server/Models/RequestModel/BillRequest.cs
@@ -10,7 +10,7 @@
 
         public Guid? PromotionId { get; set; }
 
-        public long? UsedPoint { get; set; }
+        public decimal? UsedPoint { get; set; }
 
         //public Guid? PromotionId { get; set; }
     }
